--- iz/src/main/java/com/izikgram/global/config/MyBatisConfig.java
@@ -1,9 +1,0 @@
-package com.izikgram.global.config;
-
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@MapperScan(basePackages = "com.izikgram.**.repository")
-public class MyBatisConfig {
-}
