+++ iz/src/main/java/com/izikgram/user/controller/UserController.java
@@ -1,20 +1,15 @@
 package com.izikgram.user.controller;
 
 import com.izikgram.global.config.CustomUserDetails;
-import com.izikgram.global.config.CustomUserDetailsService;
 import com.izikgram.user.entity.User;
 import com.izikgram.user.service.UserService;
-import jakarta.servlet.http.HttpSession;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -26,6 +21,7 @@ public class UserController {
 
     @Autowired
     private UserService userService;
+
     @Autowired
     private PasswordEncoder passwordEncoder;
